Rename package from lia-web to cross-web
- Rename src/lia to src/cross_web
- Update all imports to use cross_web
- Update pyproject.toml with new package name
- Bump version to 0.3.0
