+++ tests/request/test_base.py
@@ -1,4 +1,4 @@
-from lia.request._base import FormData
+from cross_web.request._base import FormData
 
 
 def test_form_data_creation() -> None:
