+++ tests/response/test_response.py
@@ -1,4 +1,4 @@
-from lia import Response
+from cross_web import Response
 
 
 def test_redirect() -> None:
